Added more styling. Added background to index.

+++ public/views/snippets/stock.controller.js
@@ -11,48 +11,63 @@ var app = angular.module('myapp', ["highcharts-ng"]);
 
 app.controller('stocks', function ($scope, $http, $timeout) {
 	$scope.addStock = function () {
-		var obj = {
-			stock: $scope.newStock.toUpperCase()
-		}
-		$http.post("/stock/new", obj).then(function (res) {
-			if (res.status == 200) {
-				if ($scope.stocks == undefined) {
-					$scope.stocks = [];
-				}
-				let objec = {
-					symbol: obj.stock
-				}
+		if ($scope.loading != true) {
+			$scope.loading = true;
+			$scope.message = "";
 
-				let elem = res.data[res.data.length - 1];
-
-				$scope.chartConfig.series.push({name: elem.symbol, data: elem.data});
-
-				$http.post("/snapshot", objec).then(function (resp) {
-					$scope.stocks.push({symbol: resp.data.symbol, name: resp.data.name});
-				});
-
-			} else {
-				console.log(res);
+			var obj = {
+				stock: $scope.newStock.toUpperCase()
 			}
-		}, function (err) {
-			console.error(err);
-		});
+			$http.post("/stock/new", obj).then(function (res) {
+				if (res.status == 200) {
+					if ($scope.stocks == undefined) {
+						$scope.stocks = [];
+					}
+					let objec = {
+						symbol: obj.stock
+					}
+
+					let elem = res.data[res.data.length - 1];
+
+					$scope.chartConfig.series.push({name: elem.symbol, data: elem.data});
+
+					$http.post("/snapshot", objec).then(function (resp) {
+						$scope.stocks.push({symbol: resp.data.symbol, name: resp.data.name});
+						$scope.loading = false;
+					});
+
+				} else {
+					$scope.message = res.data;
+					$scope.loading = false;
+				}
+			}, function (err) {
+				$scope.message = err.data;
+				$scope.loading = false;
+			});
+		}
 	}
 
 	$scope.removeStock = function (stock) {
-		$http.delete("/stock/" + stock.symbol).then(function (res) {
-			var index = findOptionIndex($scope.stocks, stock);
-			$scope.stocks.splice(index, 1);
+		if ($scope.loading != true) {
+			$scope.loading = true;
+			$scope.message = "";
 
-			if ($scope.stocks.length === 0) {
-				$scope.stocks = undefined;
-				$scope.chartConfig.series = [];
-			} else {
-				$scope.chartConfig.series.splice(index, 1);
-			}
-		}, function (err) {
-			console.error(err);
-		})
+			$http.delete("/stock/" + stock.symbol).then(function (res) {
+				var index = findOptionIndex($scope.stocks, stock);
+				$scope.stocks.splice(index, 1);
+
+				if ($scope.stocks.length === 0) {
+					$scope.stocks = undefined;
+					$scope.chartConfig.series = [];
+				} else {
+					$scope.chartConfig.series.splice(index, 1);
+				}
+				$scope.loading = false;
+			}, function (err) {
+				$scope.message = err.data;
+				$scope.loading = false;
+			})
+		}
 	}
 
 	$scope.chartConfig = {
@@ -84,6 +99,10 @@ app.controller('stocks', function ($scope, $http, $timeout) {
 		}
 	}
 
+	$scope.closeMessage = function () {
+		$scope.message = "";
+	}
+
 	<% if(user.stocks[0]) { %>
 	$scope.stocks = [
 		<% for(var i = 0; i < user.stocks.length -1; i++) { %>{
@@ -97,5 +116,7 @@ app.controller('stocks', function ($scope, $http, $timeout) {
 	]<% } %>
 
 	$scope.newStock = "";
+	$scope.loading = false;
 	$scope.setInitial($scope.stocks);
+	angular.element(document).find(".nojs").removeClass("nojs");
 });
